+++ uui/src/lib/Box/box.types.d.ts
@@ -1,8 +1,9 @@
+import type { ScopedStyles } from '$lib/shared/theme/config/theme-config.types';
 import type { current_component } from 'svelte/internal';
 
 export interface BoxProps<T extends EventTarget> extends svelte.JSX.HTMLAttributes<T> {
   element?: string;
-  ref?: HTMLElement;
-  additionalEvents?: string[];
+  ref?: Node;
   eventComponent?: typeof current_component;
+  scopedStyle?: ScopedStyles;
 }