+++ uui/src/lib/index.ts
@@ -8,11 +8,13 @@ import Dialog from './Dialog/Dialog.svelte';
 import Stepper from './Stepper/Stepper.svelte';
 import ChipIcon from './Chip/chip-icon.svelte';
 import StepIcon from './Stepper/StepIcon.svelte';
-import ButtonIcon from './Button/ButtonIcon.svelte';
+import TextField from './TextField/TextField.svelte';
 import Accordion from './Accordion/accordion.svelte';
-import DialogTitle from './Dialog/DialogTitle.svelte';
+import ButtonIcon from './Button/ButtonIcon.svelte';
 import IconButton from './IconButton/IconButton.svelte';
 import Typography from './Typography/typography.svelte';
+import DialogTitle from './Dialog/DialogTitle.svelte';
+import TextFieldIcon from './TextField/TextFieldIcon.svelte';
 import DialogActions from './Dialog/DialogActions.svelte';
 import AccordionGroup from './Accordion/accordion-group.svelte';
 
@@ -26,11 +28,13 @@ export {
   Stepper,
   ChipIcon,
   StepIcon,
+  TextField,
   Accordion,
   ButtonIcon,
   IconButton,
   Typography,
   DialogTitle,
+  TextFieldIcon,
   DialogActions,
   AccordionGroup
 };
@@ -38,6 +42,7 @@ export {
 // Export Utility Functions here
 export { makeThemeColor } from './shared/utils/makeThemeColor';
 export { getPreferredTheme } from './shared/utils/getPreferredTheme';
+export { scopedStyles } from './shared/utils/scopedStyles';
 export { UUI_PALETTES } from './shared/theme/palette/default-palettes';
 export { UUI_TYPOGRAPHY } from './shared/theme/typography-config/default-typography-config';
 
@@ -64,13 +69,17 @@ export type {
   PreferredTheme
 } from './shared/theme/palette/palette.types';
 export type { BoxProps } from './Box/box.types';
-export type { CssObject } from './shared/theme/config/theme-config.types';
+export type {
+  CssObject,
+  ScopedStyles,
+  BaseScopedStyle,
+  CssProperties
+} from './shared/theme/config/theme-config.types';
 export type {
   TypographyVars,
   ThemeTypography,
   TypographyConfig,
-  TypographyPalette,
-  TypographyPaletteValues
+  TypographyPalette
 } from './shared/theme/typography-config/typography-config.types';
 export type { DialogActionsProps, DialogTitleProps } from './Dialog/Dialog.types';
 export type { IconButtonColor, IconButtonProps } from './IconButton/IconButton.types';
@@ -81,5 +90,6 @@ export type {
   CustomStep,
   CustomSteps
 } from './Stepper/stepper.types';
+export type { TextFieldProps, TextFieldIconProps } from './TextField/TextField.types';
 export type { AccordionProps, AccordionGroupProps } from './Accordion/accordion.types';
 export type { ThemeColors } from './shared/theme/default-colors/colors.types';